requirements.txt
python-dotenv==1.0.0
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
pypdf==3.17.1
tiktoken==0.5.2
chromadb==0.4.18
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
numpy==1.26.2